Create separate files for auth's schemas and models fix unused import in security.py
