+++ app/core/security.py
@@ -2,7 +2,7 @@ import secrets
 from datetime import datetime, timedelta
 
 from fastapi.security import OAuth2PasswordBearer
-from jose import jwk, jwt
+from jose import jwt
 from passlib.context import CryptContext
 from sqlalchemy.ext.asyncio import AsyncSession
